+++ packages/navi-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @awell-health/navi-core
 
+## 0.0.6
+
+### Patch Changes
+
+- 7a4f7b0: feat(session): authenticated state is separated from session data and should exist only in the JWT
+- Beta: Updated data change activity event
+
 ## 0.0.5
 
 ### Patch Changes