+++ packages/navi-js/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @awell-health/navi-js
 
+## 0.0.6
+
+### Patch Changes
+
+- 7a4f7b0: feat(session): authenticated state is separated from session data and should exist only in the JWT
+- Beta: Updated data change activity event
+- Updated dependencies [7a4f7b0]
+- Updated dependencies
+  - @awell-health/navi-core@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes